+++ src/main/java/pl/plajer/pinata/commands/MainCommand.java
@@ -34,18 +34,15 @@ import pl.plajer.pinata.creator.CreatorMenu;
 import pl.plajer.pinata.pinata.LivingPinata;
 import pl.plajer.pinata.utils.Utils;
 
-public class MainCommand implements CommandExecutor {
+public class MainCommand{
 
   private Map<Entity, LivingPinata> pinata = new HashMap<>();
   private List<Player> users = new ArrayList<>();
-  private GameCommands argumentsManager;
   private Main plugin;
 
   public MainCommand(Main plugin, boolean register) {
     if (register) {
       this.plugin = plugin;
-      argumentsManager = new GameCommands(plugin);
-      plugin.getCommand("pinata").setExecutor(this);
     }
   }
 
@@ -61,83 +58,6 @@ public class MainCommand implements CommandExecutor {
     if (!sender.hasPermission(permission)) {
       sender.sendMessage(Utils.colorMessage("Pinata.Command.No-Permission"));
       return false;
-    }
-    return true;
-  }
-
-  @Override
-  public boolean onCommand(final CommandSender sender, Command command, String label, String[] args) {
-    if (!hasPermission(sender, "pinata.command")) {
-      return true;
-    }
-    if (args.length == 0) {
-      sender.sendMessage(Utils.colorMessage("Pinata.Command.Help-Command.Header"));
-      sender.sendMessage(Utils.colorMessage("Pinata.Command.Help-Command.Description"));
-      return true;
-    }
-    if (args[0].equalsIgnoreCase("list")) {
-      if (!isSenderPlayer(sender)) {
-        return true;
-      }
-      if (!hasPermission(sender, "pinata.command.list")) {
-        return true;
-      }
-      Utils.createPinatasGUI("Menus.List-Menu.Inventory-Name", (Player) sender);
-    } else if (args[0].equalsIgnoreCase("preview")) {
-      argumentsManager.openPreviewMenu(sender, args);
-    } else if (args[0].equalsIgnoreCase("buy")) {
-      argumentsManager.buyPinata(sender, args);
-    } else if (args[0].equalsIgnoreCase("reloadconfig")) {
-      argumentsManager.reloadConfig(sender);
-    } else if (args[0].equalsIgnoreCase("setchance")) {
-      if (!isSenderPlayer(sender)) {
-        return true;
-      }
-      if (!hasPermission(sender, "pinata.admin.setchance")) {
-        return true;
-      }
-      if (args.length == 1) {
-        sender.sendMessage("Please type chance!");
-      } else {
-        argumentsManager.applyChanceToItem(sender, args[1]);
-      }
-    } else if (args[0].equalsIgnoreCase("create")) {
-      argumentsManager.createPinata(sender, args);
-    } else if (args[0].equalsIgnoreCase("setcrate")) {
-      argumentsManager.setCrate(sender, args);
-    } else if (args[0].equalsIgnoreCase("cratelist")) {
-      argumentsManager.printCrateList(sender);
-    } else if (args[0].equalsIgnoreCase("createnew")) {
-      if (!isSenderPlayer(sender)) {
-        return true;
-      }
-      if (!hasPermission(sender, "pinata.admin.createpinata")) {
-        return true;
-      }
-      if (args.length == 1) {
-        sender.sendMessage("Please type pinata name!");
-      } else {
-        argumentsManager.createNewPinata(sender, args[1]);
-      }
-    } else if (args[0].equalsIgnoreCase("edit")) {
-      if (!isSenderPlayer(sender)) {
-        return true;
-      }
-      if (!hasPermission(sender, "pinata.admin.edit")) {
-        return true;
-      }
-      if (args.length == 1) {
-        sender.sendMessage("Please type pinata name!");
-      } else {
-        if (plugin.getPinataManager().getPinataByName(args[1]) == null) {
-          sender.sendMessage("Pinata doesn't exist!");
-          return true;
-        }
-        new CreatorMenu(args[1]).openInventory((Player) sender);
-      }
-    } else {
-      sender.sendMessage(Utils.colorMessage("Pinata.Command.Help-Command.Header"));
-      sender.sendMessage(Utils.colorMessage("Pinata.Command.Help-Command.Description"));
     }
     return true;
   }