+++ src/main/java/pl/plajer/pinata/utils/Utils.java
@@ -37,7 +37,6 @@ import org.bukkit.plugin.java.JavaPlugin;
 import pl.plajer.pinata.Main;
 import pl.plajer.pinata.api.PinataFactory;
 import pl.plajer.pinata.handlers.language.LanguageManager;
-import pl.plajer.pinata.handlers.language.Locale;
 import pl.plajer.pinata.pinata.Pinata;
 import pl.plajerlair.core.services.exception.ReportedException;
 import pl.plajerlair.core.utils.MinigameUtils;
@@ -58,11 +57,6 @@ public class Utils {
       e1.printStackTrace();
       MessageUtils.errorOccurred();
       Bukkit.getConsoleSender().sendMessage("Game message not found!");
-      if (LanguageManager.getPluginLocale() == Locale.ENGLISH) {
-        Bukkit.getConsoleSender().sendMessage("Please regenerate your language.yml file! If error still occurs report it to the developer!");
-      } else {
-        Bukkit.getConsoleSender().sendMessage("Locale message string not found! Please contact developer!");
-      }
       Bukkit.getConsoleSender().sendMessage("Access string: " + message);
       return "ERR_MESSAGE_NOT_FOUND";
     }