+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Added pinata create
 * Added pinata edit
 * Added pinata cratelist
+* Added ArgumentsRegistry
 
 ### 2.6.3 update (27.03.2018)
 * Some code changes